# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/employee_bank_accounts:
    get:
      tags:
        - EmployeeBankAccounts
      summary: List EmployeeBankAccounts
      description: Get a paginated list of EmployeeBankAccounts with optional filtering and sorting
      operationId: listEmployeeBankAccounts
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeBankAccountList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - EmployeeBankAccounts
      summary: Create EmployeeBankAccount
      description: Create a new EmployeeBankAccount
      operationId: createEmployeeBankAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmployeeBankAccountRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeBankAccount'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/employee_bpjses:
    get:
      tags:
        - EmployeeBpjses
      summary: List EmployeeBpjses
      description: Get a paginated list of EmployeeBpjses with optional filtering and sorting
      operationId: listEmployeeBpjses
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeBpjsList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - EmployeeBpjses
      summary: Create EmployeeBpjs
      description: Create a new EmployeeBpjs
      operationId: createEmployeeBpjs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmployeeBpjsRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeBpjs'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    EmployeeBankAccount:
      type: object
      required:
        - id
        - employee_id
        - bank_id
        - account_number
        - metadata
      properties:
        id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        bank_id:
          type: string
          format: uuid
        account_number:
          type: string
        account_name:
          nullable: true
          type: string
        metadata:
          type: object
...
    CreateEmployeeBankAccountRequest:
      type: object
      required:
        - employee_id
        - bank_id
        - account_number
        - metadata
      properties:
        employee_id:
          type: string
          format: uuid
        bank_id:
          type: string
          format: uuid
        account_number:
          type: string
        account_name:
          nullable: true
          type: string
        metadata:
          type: object
...
    EmployeeBankAccountList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/EmployeeBankAccount'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    EmployeeBpjs:
      type: object
      required:
        - id
        - employee_id
        - metadata
      properties:
        id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        bpjs_ketenagakerjaan_number:
          nullable: true
          type: string
        npp_bpjs_ketenagakerjaan:
          nullable: true
          type: string
        bpjs_ketenagakerjaan_date:
          nullable: true
          type: string
          format: date
        bpjs_kesehatan_number:
          nullable: true
          type: string
        bpjs_kesehatan_family:
          nullable: true
          type: integer
          format: int32
        bpjs_kesehatan_date:
          nullable: true
          type: string
          format: date
        jaminan_pensiun_date:
          nullable: true
          type: string
          format: date
        metadata:
          type: object
...
    CreateEmployeeBpjsRequest:
      type: object
      required:
        - employee_id
        - metadata
      properties:
        employee_id:
          type: string
          format: uuid
        bpjs_ketenagakerjaan_number:
          nullable: true
          type: string
        npp_bpjs_ketenagakerjaan:
          nullable: true
          type: string
        bpjs_ketenagakerjaan_date:
          nullable: true
          type: string
          format: date
        bpjs_kesehatan_number:
          nullable: true
          type: string
        bpjs_kesehatan_family:
          nullable: true
          type: integer
          format: int32
        bpjs_kesehatan_date:
          nullable: true
          type: string
          format: date
        jaminan_pensiun_date:
          nullable: true
          type: string
          format: date
        metadata:
          type: object
...
    EmployeeBpjsList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/EmployeeBpjs'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []